+++ rdagent/log/mle_summary.py
@@ -2,18 +2,19 @@ import json
 import pickle
 import re
 from collections import defaultdict
-from datetime import timedelta
 from pathlib import Path
 
 import fire
 import pandas as pd
 
-from rdagent.app.data_science.conf import DS_RD_SETTING
-from rdagent.app.data_science.loop import DataScienceRDLoop
-from rdagent.components.coder.data_science.conf import get_ds_env
 from rdagent.core.experiment import FBWorkspace
 from rdagent.core.proposal import ExperimentFeedback
 from rdagent.log.storage import FileStorage
+from rdagent.log.utils import (
+    extract_json,
+    extract_loopid_func_name,
+    is_valid_session,
+)
 from rdagent.log.utils.folder import get_first_session_file_after_duration
 from rdagent.scenarios.data_science.experiment.experiment import DSExperiment
 from rdagent.scenarios.data_science.test_eval import (
@@ -27,19 +28,6 @@ from rdagent.utils.workflow import LoopBase
 test_eval = get_test_eval()
 
 is_mle = isinstance(test_eval, MLETestEval)
-
-
-def extract_mle_json(log_content: str) -> dict | None:
-    match = re.search(r"\{.*\}", log_content, re.DOTALL)
-    if match:
-        return json.loads(match.group(0))
-    return None
-
-
-def extract_loopid_func_name(tag):
-    """提取 Loop ID 和函数名称"""
-    match = re.search(r"Loop_(\d+)\.([^.]+)", tag)
-    return match.groups() if match else (None, None)
 
 
 def save_grade_info(log_trace_path: Path):
@@ -59,10 +47,6 @@ def save_grade_info(log_trace_path: Path):
                     )
                 except Exception as e:
                     print(f"Error in {log_trace_path}: {e}")
-
-
-def is_valid_session(p: Path) -> bool:
-    return p.is_dir() and p.joinpath("__session__").exists()
 
 
 def save_all_grade_info(log_folder):
@@ -149,7 +133,7 @@ def summarize_folder(log_folder: Path, hours: int | None = None):
                             env=test_eval.env,
                             entry=f"mlebench grade-sample None {stat[log_trace_path.name]['competition']} --data-dir /mle/data",
                         )
-                        grade_output = extract_mle_json(stdout)
+                        grade_output = extract_json(stdout)
                         if grade_output:
                             bronze_threshold = grade_output["bronze_threshold"]
                             silver_threshold = grade_output["silver_threshold"]
@@ -164,7 +148,7 @@ def summarize_folder(log_folder: Path, hours: int | None = None):
                         if msg.content.result is not None:
                             valid_scores[loop_id] = msg.content.result
                     elif "mle_score" in msg.tag:
-                        grade_output = extract_mle_json(msg.content)
+                        grade_output = extract_json(msg.content)
                         if grade_output:
                             if grade_output["submission_exists"]:
                                 made_submission_num += 1
