+++ rdagent/log/ui/conf.py
@@ -12,5 +12,7 @@ class UIBasePropSetting(ExtendedBaseSettings):
 
     aide_path: str = "./aide"
 
+    amlt_path: str = "/data/share_folder_local/amlt"
+
 
 UI_SETTING = UIBasePropSetting()
